+++ packages/create-analog/template-blog/vite.config.ts
@@ -24,6 +24,7 @@ export default defineConfig(({ mode }) => ({
     environment: 'jsdom',
     setupFiles: ['src/test.ts'],
     include: ['**/*.spec.ts'],
+    reporters: ['default'],
   },
   define: {
     'import.meta.vitest': mode !== 'production',