+++ packages/nx-plugin/src/generators/app/versions/nx_17_X/versions.ts
@@ -2,10 +2,10 @@
 // dependencies
 export const V17_X_NX_DEVKIT = '^17.0.0';
 export const V17_X_NX_ANGULAR = '^17.0.0';
-export const V17_X_ANALOG_JS_CONTENT = '^0.2.0';
-export const V17_X_ANALOG_JS_ROUTER = '^0.2.0';
-export const V17_X_ANALOG_JS_TRPC = '^0.2.0';
-export const V17_X_ANALOG_JS_VITE_PLUGIN_ANGULAR = '^0.2.0';
+export const V17_X_ANALOG_JS_CONTENT = '^1.0.0-rc.0';
+export const V17_X_ANALOG_JS_ROUTER = '^1.0.0-rc.0';
+export const V17_X_ANALOG_JS_TRPC = '^1.0.0-rc.0';
+export const V17_X_ANALOG_JS_VITE_PLUGIN_ANGULAR = '^1.0.0-rc.0';
 export const V17_X_TRPC_CLIENT = '^10.25.0';
 export const V17_X_TRPC_SERVER = '^10.25.0';
 export const V17_X_ISOMORPHIC_FETCH = '^3.0.0';
@@ -21,7 +21,7 @@ export const V17_X_POSTCSS = '^8.4.5';
 export const V17_X_AUTOPREFIXER = '^10.4.0';
 
 // devDependencies
-export const V17_X_ANALOG_JS_PLATFORM = '^0.2.0';
+export const V17_X_ANALOG_JS_PLATFORM = '^1.0.0-rc.0';
 export const V17_X_NX_VITE = '^17.0.0';
 export const V17_X_NX_LINTER = '^17.0.0';
 export const V17_X_JSDOM = '^22.0.0';