+++ packages/vite-plugin-angular/src/lib/compiler-plugin.ts
@@ -8,7 +8,10 @@
 
 import type { DepOptimizationConfig } from 'vite';
 
-import { CompilerPluginOptions, JavaScriptTransformer } from './utils/devkit';
+import {
+  CompilerPluginOptions,
+  JavaScriptTransformer,
+} from './utils/devkit.js';
 
 type EsbuildOptions = NonNullable<DepOptimizationConfig['esbuildOptions']>;
 type EsbuildPlugin = NonNullable<EsbuildOptions['plugins']>[number];