+++ packages/vite-plugin-nitro/src/index.ts
@@ -1,4 +1,4 @@
-import { nitro } from './lib/vite-plugin-nitro';
-export { Options, SitemapConfig } from './lib/options';
+import { nitro } from './lib/vite-plugin-nitro.js';
+export { Options, SitemapConfig } from './lib/options.js';
 
 export default nitro;