+++ tools/publish.sh
@@ -1,10 +1,10 @@
 #!/bin/bash
-
-npm publish node_modules/@analogjs/astro-angular --access public
-npm publish node_modules/@analogjs/content --access public
-npm publish node_modules/@analogjs/platform --access public
-npm publish node_modules/@analogjs/router --access public
-npm publish node_modules/@analogjs/trpc --access public
-npm publish node_modules/@analogjs/vite-plugin-angular --access public
-npm publish node_modules/@analogjs/vite-plugin-nitro --access public
-npm publish dist/packages/create-analog
+echo "Publishing to $RELEASE_TAG"
+#npm config set //registry.npmjs.org/:_authToken=$NPM_TOKEN
+npm publish node_modules/@analogjs/astro-angular --access public --tag $RELEASE_TAG
+npm publish node_modules/@analogjs/content --access public --tag $RELEASE_TAG
+npm publish node_modules/@analogjs/platform --access public --tag $RELEASE_TAG
+npm publish node_modules/@analogjs/router --access public --tag $RELEASE_TAG
+npm publish node_modules/@analogjs/vite-plugin-angular --access public --tag $RELEASE_TAG
+npm publish node_modules/@analogjs/vite-plugin-nitro --access public --tag $RELEASE_TAG
+npm publish dist/packages/create-analog --tag $RELEASE_TAG